rrevocationsByViolation utils separated
