+++ src/components/charts/new_revocations/RevocationsByViolation/index.js
@@ -1,0 +1,1 @@
+export { default } from "./RevocationsByViolation";
